@@ -5,13 +5,14 @@ declare(strict_types=1);
 namespace Chubbyphp\Tests\Container\Unit\Exceptions;
 
 use Chubbyphp\Container\Exceptions\ExistsException;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \Chubbyphp\Container\Exceptions\ExistsException
- *
  * @internal
  */
+#[CoversClass(ExistsException::class)]
 final class ExistsExceptionTest extends TestCase
 {
     public function testConstruct(): void
@@ -22,9 +23,7 @@ final class ExistsExceptionTest extends TestCase
         new ExistsException('test', 0);
     }
 
-    /**
-     * @dataProvider provideCreateCases
-     */
+    #[DataProvider('provideCreateCases')]
     public function testCreate(string $type): void
     {
         $exception = ExistsException::create('id', $type);
