@@ -11,10 +11,7 @@ final class Parameter
      */
     public function __construct(private $parameter) {}
 
-    /**
-     * @return mixed
-     */
-    public function __invoke()
+    public function __invoke(): mixed
     {
         return $this->parameter;
     }
